requirements.txt
streamlit
google-genai
aiofiles